Rename directory, update imports
