+++ indra_network_search/templates/search_box.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Ground and check page</title>
+
+  <!-- JS libraries -->
+  <!-- JQ -->
+  <script src="https://code.jquery.com/jquery-3.3.1.min.js" integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8=" crossorigin="anonymous"></script>
+  <!-- BS 4.1.3 -->
+  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
+
+  <!-- Links -->
+  <!-- BS 4.1.3 css -->
+  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
+
+  <!-- Vue components -->
+  <script src="https://cdn.jsdelivr.net/npm/vue/dist/vue.js"></script>
+  <script src="https://bigmech.s3.amazonaws.com/indra-db/indralabvue-dev/IndralabVue.umd.min.js"></script>
+  <link href="https://bigmech.s3.amazonaws.com/indra-db/indralabvue-dev/IndralabVue.css" rel="stylesheet">
+
+  <script src="{{ url_for('static', filename='netSearch.js') }}"></script>
+</head>
+<body>
+  {# Goals for the search box:
+     - Enter an agent
+     - Button that sends a request to server to validate the input
+     - Button that grounds the input
+  #}
+
+
+</body>
+</html>
