+++ indra_network_search/templates/vue_test.html
@@ -1,0 +1,1 @@
+/home/klas/repos/vue_tests/hello/index.html
